Use optionals without the .get()
This was a less-than-ideal usage which tested for membership
and in essence does not end up being much better than passing around
nulls in terms of code quality (in fact it may be worse due
to the longer code overall). This has now been improved.

The updates here are still to be tested.